+++ src/main/kotlin/com/github/fmueller/jarvis/ai/OllamaService.kt
@@ -21,8 +21,6 @@ import java.time.Duration
 
 object OllamaService {
 
-    var port = 11434
-
     // TODO add something about selected code has higher priority than the open files or references
     private val systemPrompt = """
                     You are Jarvis, an intelligent and helpful coding assistant on the level of an expert software developer. You assist users by providing code completions, debugging tips, explanations, and suggestions in various programming languages. Your responses are clear, concise, and directly address the user's needs.
@@ -120,6 +118,15 @@ object OllamaService {
     private val client = HttpClient.newBuilder()
         .connectTimeout(Duration.ofSeconds(2))
         .build()
+
+    /**
+     * Base URL used for all requests to Ollama.
+     */
+    var host: String = "http://localhost:11434"
+        set(value) {
+            field = value
+            assistant = createAiService()
+        }
 
     private interface Assistant {
 
@@ -219,7 +226,7 @@ object OllamaService {
     fun isAvailable(): Boolean {
         return try {
             val request = HttpRequest.newBuilder()
-                .uri(URI.create("http://localhost:$port"))
+                .uri(URI.create(host))
                 .timeout(Duration.ofSeconds(2))
                 .build()
 
@@ -233,7 +240,7 @@ object OllamaService {
     private fun isModelAvailable(): Boolean {
         return try {
             val request = HttpRequest.newBuilder()
-                .uri(URI.create("http://localhost:$port/api/tags"))
+                .uri(URI.create("$host/api/tags"))
                 .timeout(Duration.ofSeconds(2))
                 .build()
 
@@ -253,7 +260,7 @@ object OllamaService {
     private fun pullModel(): String? {
         return try {
             val request = HttpRequest.newBuilder()
-                .uri(URI.create("http://localhost:$port/api/pull"))
+                .uri(URI.create("$host/api/pull"))
                 .timeout(Duration.ofSeconds(2))
                 .header("Content-Type", "application/json")
                 .POST(HttpRequest.BodyPublishers.ofString("{\"name\":\"$modelName\"}"))
@@ -306,7 +313,7 @@ object OllamaService {
             .streamingChatModel(
                 OllamaStreamingChatModel.builder()
                     .timeout(Duration.ofMinutes(5))
-                    .baseUrl("http://localhost:$port")
+                    .baseUrl(host)
                     .modelName(modelName)
                     .build()
             )